test(games): extract the game seeder into a shared test helper
blob-cache.test.ts kept makeBlob/putBlob/seedGame local to the file. A
second games/* test needs the same games row + R2 blob pair, and copying
~45 lines of seed SQL into it is how two fixtures drift into disagreeing
about what a seeded game looks like. One helper instead.

Pure move — no behavior change, blob-cache.test.ts asserts exactly what
it did before.
